A Simple Game
https://github.com/libgdx/libgdx/wiki/A-simple-game

This was a really simple tutorial that I found extremely helpful for
getting an abstract idea about how LibGDX structures the application.
The link is the full tutorial walkthrough. I tried to comment the code
well so that you can get an good idea of what's going on

It's only 2 classes and took me like an hour after I figured out getting
the projects to build without error. Fun stuff