RESET
The project was initially set up manually (rather than using the
gdx-setup-ui.jar utility). I forgot to initially setup the -android
project, which is required. I tried setting it up manually but it was
being a pain in the ass.

Since we have nothing in this project yet, its not a big deal to
recreate it, automated this time, so it'll all be correct