GameFrame 2.0
At this point, the project has been recreated correctly, the window size
is set to 1243x768 (a golden rectangle) and the screen should be a deep
blue.

Check it out by running DevMaze that's in the desktop project. I'm
currently getting errors in the HTML project, but I think it's just one
last dependency issue I don't feel like finding at the moment