Pulled changes accidentally sent to master
This is indentical to the 'Portal Groundwork' commit from the master
branch except for the tiny change you see here.

If this runs on your computer, let me know. For whatever reason, it
can't find swatch.png even though everything is completely totally right
as far as I can tell